more opr digits

+++ rankings/models.py
@@ -44,12 +44,12 @@ class TeamYear(models.Model):
     elo_mean = models.IntegerField()  # ex: 1746
     elo_max = models.IntegerField()  # ex: 1746
     elo_diff = models.IntegerField()  # ex: 100
-    opr = models.IntegerField()  # ex: 50
-    opr_auto = models.IntegerField()  # ex: 50
-    opr_teleop = models.IntegerField()  # ex: 50
-    opr_1 = models.IntegerField()  # ex: 50
-    opr_2 = models.IntegerField()  # ex: 50
-    opr_endgame = models.IntegerField()  # ex: 50
+    opr = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_auto = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_teleop = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_1 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_2 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_endgame = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
     ils_1 = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 50
     ils_2 = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 50
 
@@ -70,9 +70,9 @@ class Event(models.Model):
     elo_top8 = models.IntegerField()  # ex: 1800
     elo_top24 = models.IntegerField()  # ex: 1700
     elo_mean = models.IntegerField()  # ex: 1600
-    opr_top8 = models.IntegerField()  # ex: 50
-    opr_top24 = models.IntegerField()  # ex: 40
-    opr_mean = models.IntegerField()  # ex: 30
+    opr_top8 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_top24 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_mean = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
 
 
 class TeamEvent(models.Model):
@@ -86,15 +86,15 @@ class TeamEvent(models.Model):
     elo_mean = models.IntegerField()  # ex: 1746
     elo_max = models.IntegerField()  # ex: 1746
     elo_diff = models.IntegerField()  # ex: 100
-    opr_start = models.IntegerField()  # ex: 50
-    opr_end = models.IntegerField()  # ex: 50
-    opr_auto = models.IntegerField()  # ex: 10
-    opr_teleop = models.IntegerField()  # ex: 20
-    opr_1 = models.IntegerField()  # ex: 15
-    opr_2 = models.IntegerField()  # ex: 20
-    opr_endgame = models.IntegerField()  # ex: 20
-    opr_fouls = models.IntegerField()  # ex: 2
-    opr_no_fouls = models.IntegerField()  # ex: 48
+    opr_start = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_end = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_auto = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_teleop = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_1 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_2 = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
+    opr_endgame = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_fouls = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
+    opr_no_fouls = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34  # noqa 502
     ils_1_start = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 0.52  # noqa 502
     ils_2_start = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 0.52  # noqa 502
     ils_1_end = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 0.52
@@ -156,7 +156,7 @@ class TeamMatch(models.Model):
     match = models.CharField(max_length=10)  # ex: sf1m1
     alliance = models.CharField(max_length=10)  # ex: red
     elo = models.IntegerField()  # ex: 1746
-    opr = models.IntegerField()  # ex: 40
+    opr = models.DecimalField(max_digits=5, decimal_places=2)  # ex: 12.34
     ils_1 = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 0.61
     ils_2 = models.DecimalField(max_digits=3, decimal_places=2)  # ex: 0.51
